% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{hsl_logo}
\alias{hsl_logo}
\title{HSL Logo}
\usage{
hsl_logo(path = NULL)
}
\arguments{
\item{path}{If the rmd file is in a subdirectory, set the path here.}
}
\description{
Put the HSL logo file stored in this package in the local folder
and then return the new file path.
}
